requirements.txt
plotly
pandas
numpy
datetime
requests
netCDF4
lxml
argparse
pytz
entsoe-py
pvlib
joblib
scikit-learn
tensorflow
keras